refactor: 도메인 로직과 뷰 로직 분리 및 구조 개선
- RacingcarManager에서 InputView, OutputView, Validate 의존성 제거
- App.js에서 입력/출력 로직 및 유효성 검증 처리하도록 변경
- domain 폴더와 view 폴더로 파일 이동하여 책임 분리
- 불필요한 의존성 제거
